+++ src/components/FormPanel/FormPanel.tsx
@@ -636,6 +636,7 @@ export const FormPanel: React.FC<Props> = ({
             <tr>
               <td data-testid={TestIds.panel.singleLayoutContent}>
                 <FormElements
+                  data={data}
                   options={options}
                   elements={elements}
                   onChangeElement={onChangeElement}
@@ -655,6 +656,7 @@ export const FormPanel: React.FC<Props> = ({
                     <td className={styles.td} key={id} data-testid={TestIds.panel.splitLayoutContent(section.name)}>
                       <FieldSet label={section.name}>
                         <FormElements
+                          data={data}
                           options={options}
                           elements={elements}
                           onChangeElement={onChangeElement}
@@ -684,6 +686,7 @@ export const FormPanel: React.FC<Props> = ({
                             initial={initial}
                             section={section}
                             replaceVariables={replaceVariables}
+                            data={data}
                           />
                         </FieldSet>
                       </td>