+++ src/constants/form-element.ts
@@ -156,3 +156,27 @@ export const IconOptions = getAvailableIcons().map((name) => ({
   value: name,
   icon: name,
 }));
+
+/**
+ * Options Source
+ */
+export enum OptionsSource {
+  Query = 'Query',
+  Custom = 'Custom',
+}
+
+/**
+ * Options Source Options
+ */
+export const OptionsSourceOptions = [
+  {
+    label: 'Query',
+    value: OptionsSource.Query,
+    ariaLabel: TestIds.formElementsEditor.optionsSourceOption(OptionsSource.Query),
+  },
+  {
+    label: 'Custom',
+    value: OptionsSource.Custom,
+    ariaLabel: TestIds.formElementsEditor.optionsSourceOption(OptionsSource.Custom),
+  },
+];